contest {rcvr}	R Documentation

Create contest object

Description

Constructor for instances of S3 class contest implemented as tbl_df with metadata.

Usage

contest(
  contest_name,
  source = NULL,
  CVR_files = NULL,
  original = FALSE,
  contest_str = NULL,
  date_str = NULL,
  office = NULL,
  jurisdiction = NULL,
  alternatives = NULL,
  ambiguities = NULL,
  creator = NULL,
  quietly = TRUE
)

Arguments

contest_name
For example "Democratic Primary". With padding removed, used for first part of the *.rda file name for serializing the contest.

source
Folder of CVR files.

CVR_files
Cast Vote Record file names.

original
If original == TRUE, contest_str must match an object value in contest_keys.txt file in working directory, with cvr_url value and all cvr_file values needed to reconstruct the election.

contest_str
Must reference an object value in key_values.txt.

date_str
Encoded as MM/DD/YYYY. Forms the second part of the *.rda file name for serializing the contest.

office
Encodes the name of the office or referendum at issue. With padding removed, forms the third part of the *.rda file name for serializing the contest.

jurisdiction
Set as appropriate.

alternatives
Container for all alternatives in a contest (char).

ambiguities
Container for ambiguities in a contest.

creator
Human creating contest object.

quietly
When FALSE many routines get noisy.

Value

An instance of class contest.

[Package rcvr version 0.1.0 Index]