encode_cvr_rankings {rcvr}	R Documentation

Encode cast vote record rankings

Description

Encodes rankings on each ranked choice cvr as a single string rank_str with a character [A-Z] for each ranking, designating a specific alternative_code or ambiguity_code.

Usage

encode_cvr_rankings(cntst, cvr_mark_code)

Arguments

cntst
A contest instance.

cvr_mark_code
The concatenation of alternative_code and ambiguity_code. The vector of characters to used to encode each ballot. Alternatives use as.character(LETTERS[1:num_alternatives]) with U and O for alternatives. names(cvr_mark_code) corresponds to entries in the choice fields of cntst.

Value

A rank_str) derived from ranking data for each record.

[Package rcvr version 0.1.0 Index]