params {rcvr}	R Documentation

rcvr parameters

Description

The rcvr package provides functions for reading, analyzing, and visualizing ballot data from Ranked Choice Voting elections.

When rcvr is installed, missing *url* arguments default to the *extdata* subfolder of the installed *rcvr* package folder on the local machine (*system.file("extdata", package = "rcvr"). If space is an issue, the utility *dump_default_files()* is provided for your cobvenience.

For ease of of development, second copy of default data files are installed the *extdata* subfolder of the *rcvr* package development directory (*system.file("extdata", package = "rcvr*).

Arguments

alternative
Candidate or alternative for a contest (char).

alternatives
Container for all alternatives in a contest (char).

alternative_code
Character vector of codes "[A-N]" for representing alternatives in a contest.

ambiguity
The string found on the CVR representing an unresolved ranking.

ambiguities
Container for ambiguities in a contest.

contest_name
For example "Democratic Primary". With padding removed, used for first part of the *.rda file name for serializing the contest.

contest_str
Must reference an object value in key_values.txt.

continuing_alternatives
A subset of alternative_code representing continuing alternatives in a tabulation.

cntst
A contest instance.

creator
Human creating contest object.

CVR
A list of dataframes with cast vote records. Usually raw data saved used by the constructor and saved as metadata for a contest instance.

CVR_files
Cast Vote Record file names.

date_str
Encoded as MM/DD/YYYY. Forms the second part of the *.rda file name for serializing the contest.

june2018
If TRUE duplicates official State tabulations.

jurisdiction
Set as appropriate.

source
Folder of CVR files.

office
Encodes the name of the office or referendum at issue. With padding removed, forms the third part of the *.rda file name for serializing the contest.

original
If original == TRUE, contest_str must match an object value in contest_keys.txt file in working directory, with cvr_url value and all cvr_file values needed to reconstruct the election.

path
Some folder, somewhere. Typically defaults to working directory.

quietly
When FALSE many routines get noisy.

round
An instance of a RC/IR runoff round.

url
Universal Resource Locator.

[Package rcvr version 0.1.0 Index]