tabulate_cvr_rank {rcvr}	R Documentation

Tabulate CVR rank string

Description

Implements rules for the disposition of a cast vote record rank_str, including the disposition of ambiguities such as overvotes and undervotes.

Usage

tabulate_cvr_rank(
  rank_str,
  continuing_alternatives,
  sequentialO = FALSE,
  max_O = 1,
  sequentialU = FALSE,
  max_U = 2,
  june2018 = FALSE
)

Arguments

sequentialO
If TRUE, max_O sequential overvotes exhaust the cvr by undervote, status_code = "exo".

max_O
The number of overvotes triggering a potential overvote, according to the value of sequentialO.

sequentialU
If TRUE, max_u sequential undervotes exhaust the cvr by undervote, status_code = "exu".

max_U
The number of undervotes triggering a potential undervote, according to the value of sequentialO.

Value

A value from status_code

[Package rcvr version 0.1.0 Index]