tabulate_round {rcvr}	R Documentation

Tabulate instant runoff round

Description

Tabulates the result of an instant runoff round for a ranked choice contest.

Usage

tabulate_round(cntst, result, cvr_status_code, round_number)

Arguments

result
A list of results from tabulate_cvr_rank.

cvr_status_code
Codes used encode disposition of a cvr during an instant runoff round.

round_number
The runoff round number.

Value

A tabulation instance.

[Package rcvr version 0.1.0 Index]